# docs/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "code": { "example": "AUTHORIZATION_DENIED", "type": "string" },
          "correlationId": { "format": "uuid", "type": "string" },
          "message": {
            "oneOf": [
              { "type": "string" },
              { "items": { "type": "string" }, "type": "array" }
            ]
          },
          "path": {
            "example": "/api/v1/organizations/30000000-0000-4000-8000-000000000001/plants",
            "type": "string"
          },
          "statusCode": { "example": 403, "type": "integer" },
          "timestamp": { "format": "date-time", "type": "string" }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "path",
          "timestamp",
          "correlationId"
        ],
        "type": "object"
      },
...
      "PinDto": {
        "properties": {
          "pin": {
            "maxLength": 12,
            "minLength": 6,
            "pattern": "^\\d+$",
            "type": "string"
          }
        },
        "required": ["pin"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "supabase": {
        "bearerFormat": "Supabase access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Contrato backend del sistema de planta. Todos los ejemplos son ficticios.",
    "title": "Industrias Doradas API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/{organizationId}/stations/{stationId}/elevations": {
      "post": {
        "operationId": "StationController_elevate",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "stationId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PinDto" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": { "schema": { "type": "object" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Station"]
      }
    },
    "/api/v1/organizations/{organizationId}/stations/{stationId}/session-snapshot": {
      "get": {
        "operationId": "StationController_snapshot",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "stationId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "object" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Station"]
      }
    },
...
    "/api/v1/profile/pin": {
      "post": {
        "operationId": "PinController_set",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PinDto" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": { "schema": { "type": "object" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Pin"]
      }
    },
    "/api/v1/profile/pin/reset-blocks": {
      "post": {
        "operationId": "PinController_reset",
        "parameters": [],
        "responses": {
          "201": {
            "content": {
              "application/json": { "schema": { "type": "object" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Pin"]
      }
    }
  },
  "servers": [],
...
}